Add game views and remove unnecessary files - Implement MenuView, GameView, GameCompletedView, and GameOverView. - Remove .DS_Store files from the project.
